Uploaded starting files for project from local
Started this project before learning about how to use git. I know
this makes for an awkward starting point but I would rather do it
this way than "fake" going through various commits. The idea behind
this site is to have a place to showcase various screenshots that I
have taken on my PS5. Currently this is set up with a home page with
links to each game, with both a navigation bar and a slideshow preview
with links.